fix: moved TextAreaProps to types folder
